split off history from saved lib
